+++ lib/image/service/image_crop_service.dart
@@ -8,14 +8,13 @@ import 'package:kommon/image/image_crop_screen.dart';
 /// Date: 2021/8/26
 /// Mail: me@kingtous.cn
 
-class ImageCropService extends GetxService{
-
+class ImageCropService extends GetxService {
   Future<ImageCropService> init() async {
     return this;
   }
 
   Future<Uint8List?> cropImageWithFile(File file) async {
-    if (await file.exists()){
+    if (await file.exists()) {
       return null;
     }
     final path = file.path;
@@ -23,7 +22,6 @@ class ImageCropService extends GetxService{
   }
 
   Future<Uint8List?> cropImageWithPath(String path) async {
-    return await Get.to(()=>ImageCropScreen(path: path));
+    return await Get.to(() => ImageCropScreen(path: path));
   }
-
 }